requirements.txt
jinja2>=3.0.0
pyyaml>=5.4.0
ansible>=6.0.0